Conflicts:
	api/src/org/apache/cloudstack/api/BaseCmd.java
	client/tomcatconf/applicationContext.xml.in
	client/tomcatconf/componentContext.xml.in
	server/src/com/cloud/api/ApiDBUtils.java
	server/src/com/cloud/server/ManagementServerImpl.java
	server/src/com/cloud/user/AccountManagerImpl.java
	setup/db/db/schema-410to420.sql

+++ server/src/com/cloud/network/ExternalFirewallDeviceManagerImpl.java
@@ -483,7 +483,7 @@ public abstract class ExternalFirewallDeviceManagerImpl extends AdapterBase impl
 
         if (add && (!reservedIpAddressesForGuestNetwork.contains(network.getGateway()))) {
             // Insert a new NIC for this guest network to reserve the gateway address
-            _networkMgr.savePlaceholderNic(network,  network.getGateway());
+            _networkMgr.savePlaceholderNic(network,  network.getGateway(), null);
         }
         
         // Delete any mappings used for inline external load balancers in this network
@@ -501,7 +501,7 @@ public abstract class ExternalFirewallDeviceManagerImpl extends AdapterBase impl
         if (!add) {
             List<NicVO> nics = _nicDao.listByNetworkId(network.getId());
             for (NicVO nic : nics) {
-                if (nic.getReservationStrategy().equals(ReservationStrategy.PlaceHolder) && nic.getIp4Address().equals(network.getGateway())) {
+                if (nic.getVmType() == null && nic.getReservationStrategy().equals(ReservationStrategy.PlaceHolder) && nic.getIp4Address().equals(network.getGateway())) {
                     s_logger.debug("Removing placeholder nic " + nic + " for the network " + network);
                     _nicDao.remove(nic.getId());
                 }
